moved `helpers.ts` to its own folder & moved `ValidField` logic to helpers module
